Cleanup

+++ platform.h
@@ -43,7 +43,7 @@ typedef struct {
     uint32_t height;
 } PlatformTexture;
 
-static void *platformAllocate(uint64_t size);
+static void *platformAllocate(size_t size);
 
 static bool platformReadEntireFile(char *path, BaseArena *arena, BaseReadResult *readResult);
